first pass at moving key functionality out of scripts
