+++ resource_estimation/data/analysis_test.py
@@ -18,7 +18,7 @@ import cirq
 import numpy as np
 import pytest
 
-import resource_estimation.analysis as analysis
+import resource_estimation.data.analysis as analysis
 import resource_estimation.ftqc.architecture as arc
